tcg: Don't constant propagate through registers
This is to avoid retranslating blocks if the previous constant propagtion
no longer is valid. Consider a simple loop with an induction variable
initialized to 0. Upon first translation the 0 will be propagated to
uses of the induction variable. However on subsequent branches to this
block it would be retranslated and reoptimized, as the induction variable
is no longer constant.

This is not ideal for static analyses, so disable constant propagation
through registers, we still retain simpler constant propagation through
temporaries.